requirements.txt
contourpy==1.3.3
corner==2.2.3
cycler==0.12.1
emcee==3.1.6
fonttools==4.59.2
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyVirtualDisplay==3.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tzdata==2025.2
